0217 19:24

+++ pkg/errno/errno.go
@@ -7,7 +7,7 @@ import (
 )
 
 type ErrNo struct {
-	ErrCode int64
+	ErrCode int32
 	ErrMsg  string
 }
 
@@ -15,7 +15,7 @@ func (e ErrNo) Error() string {
 	return fmt.Sprintf("err_code=%d, err_msg=%s", e.ErrCode, e.ErrMsg)
 }
 
-func NewErrNo(code int64, msg string) ErrNo {
+func NewErrNo(code int32, msg string) ErrNo {
 	return ErrNo{
 		ErrCode: code,
 		ErrMsg:  msg,
@@ -28,12 +28,12 @@ func (e ErrNo) WithMessage(msg string) ErrNo {
 }
 
 var (
-	Success                = NewErrNo(int64(douyinuser.ErrCode_SuccessCode), "Success")
-	ServiceErr             = NewErrNo(int64(douyinuser.ErrCode_ServiceErrCode), "Service is unable to start successfully")
-	ParamErr               = NewErrNo(int64(douyinuser.ErrCode_ParamErrCode), "Wrong Parameter has been given")
-	UserAlreadyExistErr    = NewErrNo(int64(douyinuser.ErrCode_UserAlreadyExistErrCode), "User already exists")
-	AuthorizationFailedErr = NewErrNo(int64(douyinuser.ErrCode_AuthorizationFailedErrCode), "Authorization failed")
-	RedisServiceErr        = NewErrNo(int64(10010), "Redis service failed")
+	Success                = NewErrNo(int32(douyinuser.ErrCode_SuccessCode), "Success")
+	ServiceErr             = NewErrNo(int32(douyinuser.ErrCode_ServiceErrCode), "Service is unable to start successfully")
+	ParamErr               = NewErrNo(int32(douyinuser.ErrCode_ParamErrCode), "Wrong Parameter has been given")
+	UserAlreadyExistErr    = NewErrNo(int32(douyinuser.ErrCode_UserAlreadyExistErrCode), "User already exists")
+	AuthorizationFailedErr = NewErrNo(int32(douyinuser.ErrCode_AuthorizationFailedErrCode), "Authorization failed")
+	RedisServiceErr        = NewErrNo(int32(10010), "Redis service failed")
 )
 
 func ConvertErr(err error) ErrNo {
